Type truncation warning fixes for some RenderCore and RHI headers #codereview Rolando.Caloca #jira UE-116218
[CL 17578876 by Michael Noland in ue5-main branch]

+++ Engine/Source/Runtime/RenderCore/Public/RendererInterface.h
@@ -197,7 +197,7 @@ public:
 		ETextureCreateFlags InFlags,
 		ETextureCreateFlags InTargetableFlags,
 		bool bInForceSeparateTargetAndShaderResource,
-		uint32 InArraySize = 1,
+		uint16 InArraySize = 1,
 		uint16 InNumMips = 1,
 		bool InAutowritable = true)
 	{
@@ -235,7 +235,7 @@ public:
 		ETextureCreateFlags InFlags,
 		ETextureCreateFlags InTargetableFlags,
 		bool bInForceSeparateTargetAndShaderResource,
-		uint32 InArraySize,
+		uint16 InArraySize,
 		uint16 InNumMips = 1,
 		bool InAutowritable = true)
 	{
